Add organization and teams clients
This add for this particular stash implementation the clients to
communicate with the fowwloning endpoints:

- /rest/api/1.0/projects
- /rest/api/1.0/projects/{projectKey}
- /rest/api/1.0/projects/{projectKey}/permissions/groups
- /rest/api/1.0/projects/{projectKey}/permissions/users
- /rest/api/1.0/admin/groups
- /rest/api/1.0/admin/groups/more-members
- /rest/api/1.0/users
- /rest/api/1.0/users/{userSlug}

It used here the current implementation to list orgs(stash projects),
 groups(stash teams) and users

A small change to http client as well to hide the dependency to
retryablehttp.

The missing implementations return "not implemented" for the moment as
agreed.

Signed-off-by: Soule BA <soule@weave.works>