update logr to 1.2.4 and fix Stash client
In version 1.2.4 a [change](https://github.com/go-logr/logr/pull/166)
was introduced that significantly changed the assumptions about logr's
API, i.e. that `logr.Logger{} != logr.Discard()`. This condition
doesn't evaluate to true with 1.2.4, anymore, and the default value of
the `logr.Logger` type is now usable as a replacement for
`logr.Discard()`.

Therefore we do not need to check if the value has been provided,
anymore: If it is not provided, the default one is used which is
effectively a discarding logger.

closes #209

Signed-off-by: Max Jonas Werner <mail@makk.es>