Prevent GitLab e2e tests from running twice on PRs
The tests are already triggered by the `pull-request` event (which is
also emitted when the head branch of the PR is updated). For PRs
created from branches of the ggp repo, having the `push` event for all
branches in the workflow definition will trigger it twice. Therefore I
changed it to only run on pushes to `main` now, following what we do
in other repos.

Signed-off-by: Max Jonas Werner <mail@makk.es>