Fix Gitea tests
The Gitea tests were failing because the examples in the `./gitea`
package were run against the Gitea instance hosted at gitea.com, which
unfortunately isn't compatible with its own SDK:

```
$ curl https://gitea.com/api/v1/version
{"version":"d64e081bf6"}
```

This led to the tests to fail with this error:

```
2025/01/03 14:44:48 failed to create Gitea client: unknown version: d64e081bf6
```

I decided it's not worth it trying to make the tests run against the
locally-running e2e instance as these are run as unit tests and the
instance might not even be running and would also need to prepared
with a repository. Also, the `// Output` comment isn't very flexible
and can only contain a static string.

However, to make debugging errors in the client creation easier in the
future, I added the base URL as output to the error message.

Signed-off-by: Max Jonas Werner <max@coppersoft.eu>